requirements.txt
requests==2.32.5
apache-airflow==2.10.5